merge

+++ app/Http/Livewire/EditPermission.php
@@ -5,6 +5,8 @@ namespace App\Http\Livewire;
 use Livewire\Component;
 use App\Models\Permission;
 use App\Models\Role;
+use Illuminate\Support\Facades\Gate;
+
 class EditPermission extends Component
 {
     const route = [
@@ -13,7 +15,13 @@ class EditPermission extends Component
         'name' => 'edit-permission',
     ];
     const permission = [
-        'edit-permission' =>[
+        'view' => [
+            'title' => 'Xem quyền',
+            'description' => 'Xem quyền của vai trò',
+            'name' => 'view-permission',
+            'permission' => ['admin', 'officer'],
+        ],
+        'edit-permission' => [
             'title' => 'Cập nhật quyền',
             'description' => 'Cập nhật quyền cho vai trò',
             'name' => 'edit-permission',
@@ -22,23 +30,35 @@ class EditPermission extends Component
     ];
     public $permission;
     public $role;
-    public function mount(){
+    public function mount()
+    {
         $this->role = Role::all();
         $this->permission = Permission::all();
     }
     public function render()
     {
-        return view('livewire.admin.edit-permission');
+        if(Gate::allows('view-permission')){
+            return view('livewire.admin.edit-permission');
+        }
+        else{
+            return view('errors.not_permission');
+        }
     }
 
-    public function updatePermission($role_id, $permission_id){
-        $role = Role::find($role_id);
-        $permission = Permission::find($permission_id);
-        if($role->hasPermission($permission->name)){
-            $role->permissions()->detach($permission->id);
-        }else{
-            $role->permissions()->attach($permission->id);
+    public function updatePermission($role_id, $permission_id)
+    {
+        if (Gate::allows('edit-permission')) {
+            $role = Role::find($role_id);
+            $permission = Permission::find($permission_id);
+            if ($role->hasPermission($permission->name)) {
+                $role->permissions()->detach($permission->id);
+            } else {
+                $role->permissions()->attach($permission->id);
+            }
+            $this->dispatchBrowserEvent('alert', ['type' => 'success',  'message' => 'Cập nhật thành công!']);
         }
-        $this->dispatchBrowserEvent('alert', ['type' => 'success',  'message' => 'Cập nhật thành công!']);
+        else{
+            $this->dispatchBrowserEvent('alert', ['type' => 'warning',  'message' => 'Bạn không có quyền cập nhật!']);
+        }
     }
 }
